pyproject.toml
[project]
name = "agentic-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "optuna>=4.7.0",
    "pandas>=3.0.1",
    "pandas-market-calendars>=5.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.47",
    "yfinance>=1.2.0",
]

[tool.setuptools]
packages = ["core", "storage", "data", "broker", "backtest", "agent", "scripts", "config"]

[tool.uv.sources]
